test: de-duplicate the transport and PNTS admission tests
SonarCloud flagged duplicated lines on new code. The bodyCancelFetch fixture was
byte-identical in the EPT and 3D Tiles transport suites; it moves to
tests/helpers/bodyCancelFetch.ts and both import it. The two allocation-spy
blocks in the PNTS ceiling suite, which swap the global Float32Array and
Uint8Array to count large allocations, collapse into one countBigAllocations
helper. Behaviour is unchanged: the same 63 tests pass and typecheck is clean.
